% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{calcExactIntegralC}
\alias{calcExactIntegralC}
\title{Analytically normalizes density to one}
\description{
\code{calcExactIntegralC} is C function called by \code{
   callCalcExactIntegralC}. There you can find more details about this function.
}
