% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{calcKernelDens}
\alias{calcKernelDens}
\title{Evaluates a Kernel Density with Gaussian Kernel [C Function]}
\description{
\code{calcKernelDens} is a C function that calculates the kernel density
  with Gaussian kernel for a dataset X. It is used in the R function
  \code{paramFitKernelDensity}.
}
