% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{callAVXInfo}
\alias{callAVXInfo}
\title{Checks if AVX is active}
\usage{
callAVXInfo()
}
\description{
\code{callAVXInfo} is a wrapper to a C function, which prints information
  about whether AVX extensions are currently activated. This leads to much fuch faster
  run times. More informations about how AVX can be enabled can be found in the vignette
  about installation.
}
