% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{compilationInfo}
\alias{compilationInfo}
\title{Checks if AVX is active}
\usage{
compilationInfo()
}
\description{
\code{compilationInfo} is a wrapper to a C function, which prints information
  about whether AVX extensions where activated during compilation. Active AVX speeds up
  computations significantly (by a factor of 4 to 8 roughly). More informations about how
  AVX can be enabled during installation can be found in documentation.
}
