% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{newtonBFGSLC}
\alias{newtonBFGSLC}
\title{Optimize MLE objective for a log-concave density [C function]}
\description{
\code{newtonBFGSLC} is a C function that can be accessed by the R wrapper 
  function \code{callNewtonBFGSLC}. Check its documention for more
  details about this C function.
}
