% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{newtonBFGSLInitC}
\alias{newtonBFGSLInitC}
\title{Optimize MLE objective for smooth a log-concave density [C function]}
\description{
\code{newtonBFGSLInitC} is a C function that can be accessed by the R wrapper
  function \code{callNewtonBFGSLInitC}. Check its documention for more
  details about this C function.
} 

