% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{printAVXInfo}
\alias{printAVXInfo}
\title{Checks if AVX is active [C Function]}
\description{
\code{printAVXInfo} is a C function that can be accessed by the R wrapper function 
  \code{callAVXinfo}.
}
