% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callerFunctions.R
\name{recalcParamsC}
\alias{recalcParamsC}
\title{Calculates hyperplane parameters for the upper convex hull of conv(X,y)}
\description{
\code{recalcParamsC} is a C function that is not supposed to be called by
  the user directly.
}
